Add AnimeKai to AnimeKaiTheme and polish code
