+++ lib-multisrc/animekaitheme/build.gradle.kts
@@ -3,3 +3,8 @@ plugins {
 }
 
 baseVersionCode = 1
+
+dependencies {
+    implementation(project(":lib:megaupextractor"))
+    implementation(project(":lib:playlistutils"))
+}
