+++ src/en/animekai/build.gradle
@@ -1,11 +1,13 @@
 ext {
     extName = 'AnimeKai'
     extClass = '.AnimeKai'
-    extVersionCode = 14
+    themePkg = 'animekaitheme'
+    baseUrl = 'https://animekai.to'
+    overrideVersionCode = 14
 }
 
 apply from: "$rootDir/common.gradle"
 
 dependencies {
-    implementation(project(':lib:playlistutils'))
+    implementation(project(":lib:playlistutils"))
 }
